Add DrawingShapes

+++ InterfaceContract/Drawable.cs
@@ -7,6 +7,6 @@ namespace NeoDrawable
 {
     public interface Drawable
     {
-        List<Point2D> Draw(int startX, int startY, int endX, int endY);
+        List<Point2D> Draw(double startX, double startY, double endX, double endY);
     }
 }
